% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGridXY.R
\name{getGridXY}
\alias{getGridXY}
\title{Get x and y coordinates in a grid}
\usage{
getGridXY(n, nrows, ncols = NULL, byrow = TRUE)
}
\arguments{
\item{n}{number of coordinates to generate}

\item{nrows}{number of rows in the grid}

\item{ncols}{number of columns in the grid}

\item{byrow}{logical value indicating whether the order of coordinates in the grid is byrow or not}
}
\value{
a matrix which gives the x and y coordinates of a grid
}
\description{
getGridXY generates a matrix of x and y coordinates in a grid. The justification of each coordinate is in center.
}
\examples{
x <- getGridXY(16)
x
}
\author{
Jiahua Liu
}
