% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeanoHilbert_combine.R
\name{HilbertGlyph}
\alias{HilbertGlyph}
\title{Hilbert curve data visualization}
\usage{
HilbertGlyph(col, fill = FALSE, type = c("png", "pixmap"))
}
\arguments{
\item{col}{vector of colors}

\item{fill}{logical value indicating whether the color strings repeat from the beginning if number of colors in col is smaller than the total length of Hilbert curve}

\item{type}{string specifying the format of output. "png" means the output is a data matrix in the png format, and "pixmap" means the output is a class of pixmap.}
}
\value{
a data matrix of png format or a class of "pixmap" that is used to plot. See \code{\link[pixmap]{pixmap}}, \code{\link[png]{writePNG}}, \code{\link[graphics]{rasterImage}}
}
\description{
HilbertGlyph generates a data matrix of png format or a class of "pixmap" which can be used to display color strings in an order of Hilbert curve in the space given a vector of colors
}
\examples{
myPngmat <- HilbertGlyph()
myPngmat <- HilbertGlyph(col = colors())
myPngmat <- HilbertGlyph(col = colors(), fill = TRUE)
writePNG(myPngmat, target = "myplot.png")
plot(0,type='n', xlim=c(0,1), ylim=c(0,1), axes = FALSE,xlab = "", ylab = "")
rasterImage(myPngmat, 0, 0, 1, 1)
mypixmap <- HilbertGlyph(type = "pixmap")
mypixmap <- HilbertGlyph(col = colors(), fill = TRUE, type = "pixmap")
plot(mypixmap)

}
\author{
Jiahua Liu, Wayne Oldford
}
