% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsi2rgb.R
\name{hsi2rgb}
\alias{hsi2rgb}
\title{HSI to RGB Conversion}
\usage{
hsi2rgb(h, s = NULL, i = NULL, maxColorValue = 255)
}
\arguments{
\item{h}{vector of 'hue' values in [0,360] or 3-row hsi matrix}

\item{s}{vector of 'saturation' values in [0,1], or NULL when r is a matrix}

\item{i}{vector of 'intensity' values in [0,1], or NULL when r is a matrix}
}
\value{
A matrix with a column for each color. The three rows of the matrix indicate "red","green","blue" values and are named "r", "g", and "b" in [0,255]
        accordingly.
}
\description{
hsi2rgb transforms colors from HSI space (hue/saturation/intensity) into RGB space (red/green/blue)
}
\details{
HSI is a variation of the HSV model, which provides color scales with monotonically increasing or decreasing brightness after linear interpolation
}
\examples{
x=c(0,360)
y=c(1,0.4)
f=approxfun(x,y) # linear introploation function
l <- 100000 # number of colors
H=seq(x[1],x[2],length.out = l)
I=f(H)
S=rep(1,length(H))
R=hsi2rgb(H,S,I)[1,]
G=hsi2rgb(H,S,I)[2,]
B=hsi2rgb(H,S,I)[3,]
col1=rgb(R,G,B,maxColorValue = 255)
gr1=(0.34*R+0.5*G+0.16*B)/255
col2=grey(gr1)
par(mfrow=c(2,1))
barplot(rep(1,length(H)),col = col1,border = NA,beside = FALSE,space = c(0,0),main="HSI color mapping")
barplot(rep(1,length(H)),col = col2,border = NA,beside = FALSE,space = c(0,0))

}
\references{
Keim, Daniel A, and Hans-Peter Kriegel. 1995. "Issues in Visualizing Large Databases." In \emph{Visual Database Systems} 3, 203-14. Springer.
}
\author{
Jiahua Liu
}
