% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_glyphs_draw.R
\name{make_glyphs_draw}
\alias{make_glyphs_draw}
\title{Make glyphs in different classes}
\usage{
make_glyphs_draw(
  data,
  draw_fun,
  type = c("png", "jpeg", "pixmap", "tiff"),
  width = 50,
  height = 50,
  mar = rep(1, 4),
  ...
)
}
\arguments{
\item{data}{a list of data vectors}

\item{draw_fun}{function of drawing in a device given a data vector}

\item{type}{string specifying the format of output. "png" means the output is a list of data matrices in the png format, "jpeg" means the output is in the jpeg format and so on.}

\item{width}{the width of the device}

\item{height}{the height of the device}

\item{mar}{A numerical vector of the form c(bottom, left, top, right) which gives the number of lines of margin to be specified on the four sides of the plot}

\item{...}{Arguments passed to `png`, `jpeg`, `pixmap`, or `tiff` according to the `type`}
}
\value{
a list of glyphs in different classes corresponding to the list of data
}
\description{
Given a list of data vectors, make a corresponding list of glyphs in one of four different classes which are "png", "jpeg", "pixmap" and "tiff".
}
\examples{
n <- 16
data <- list()
for (i in 1:n){
  data[[i]] <- rnorm(500)
}
glyphs <- make_glyphs_draw(data, hist, type = "jpeg", mar = rep(4, 4), width = 200, height = 200)
x <- getGridXY(n)
plot_glyphs(x, glyphs = glyphs, glyphWidth = 0.8, glyphHeight = 0.6, axes = FALSE, xlab = "", ylab = "")


}
\author{
Wayne Oldford, Jiahua Liu
}
